+++ Persistence/database/accounts_banks_logs/structure.sql
@@ -7,9 +7,9 @@ CREATE TABLE `accounts_banks_logs` (
     `from_accounts_id`  int       NOT NULL,
     `to_accounts_id`    int       NOT NULL,
     `date`              timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    `type`              int       NOT NULL DEFAULT 1                  COMMENT 'Log type. 1 = withdraw, 2 = deposit, 3 = donation.',
+    `type`              tinyint   NOT NULL DEFAULT 1                  COMMENT 'Log type. 1 = withdraw, 2 = deposit, 3 = donation.',
     `amount`            int       NOT NULL DEFAULT 0                  COMMENT 'Amount of currency logged.',
-    `currency`          int       NOT NULL DEFAULT 1                  COMMENT 'Currency of the amount. 1 = credits, 2 = uridium.',
+    `currency`          tinyint   NOT NULL DEFAULT 1                  COMMENT 'Currency of the amount. 1 = credits, 2 = uridium.',
     `accounts_banks_id` int       NULL     DEFAULT NULL,
 
     CONSTRAINT `accounts_banks_logs_pk` PRIMARY KEY (`id`)
