+++ Persistence/database/galaxygates_gg_waves/structure.sql
@@ -3,9 +3,9 @@
 -- Many to many relations for galaxygates and galaxygates_waves.
 --
 CREATE TABLE `galaxygates_gg_waves` (
-    `id`                   int NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
-    `galaxygates_id`       int NOT NULL,
-    `galaxygates_waves_id` int NOT NULL,
+    `id`                   smallint NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
+    `galaxygates_id`       tinyint  NOT NULL,
+    `galaxygates_waves_id` tinyint  NOT NULL,
 
     CONSTRAINT `galaxygates_gg_waves_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB CHARACTER SET utf8 COMMENT 'Many to many relations for galaxygates and galaxygates_waves.';
