+++ Persistence/database/invitation_codes/structure.sql
@@ -3,9 +3,9 @@
 -- Contains the invitation codes that can be used for registering.
 --
 CREATE TABLE `invitation_codes` (
-    `id`    int           NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
+    `id`    smallint      NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
     `code`  varchar(255)  NOT NULL DEFAULT ''     COMMENT 'The invitation code.',
-    `limit` int           NOT NULL DEFAULT 1      COMMENT 'Amount of times the code can be used.',
+    `limit` tinyint       NOT NULL DEFAULT 1      COMMENT 'Amount of times the code can be used.',
 
     CONSTRAINT `invitation_codes_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB CHARACTER SET utf8 COMMENT 'Contains the invitation codes that can be used for registering.';
